X00T: Init a script that imports and sets ASUS-specific prop
Required to flash stock firmware without removing SKU and/or firmware
version assertion lines from updater-script. Be aware that you need to
be on stock firmware at least once OR before flashing recovery, so the
required props can be exported and made available on /persist for future
boot ups.

The script has to be executed late at boot, however since I've no idea
at which phase it should be; do it when adbd is restarted instead. Weird
flex, but at least it works.

Signed-off-by: Albert I <kras@raphielgang.org>